fixup! Fix background data clobbering other policies
Fetching the state of policies was not working due to an incorrect use
of Arrays.asList() - if you give it an int array, you just get a list
with that int[] as a single element, so using the contains method with
an int will always return false. This has now all been refactored.

Test: Manual: Open Settings > Network & Internet > Data Saver >
Unrestricted data. Turn it on for something. Go back, and then
return to the page. It should still show as on.

Issue: calyxos#2547
Change-Id: I671544f6fdf9897484c6265c31c8b3cd29ad4a92
Signed-off-by: Jackeagle <jackeagle102@gmail.com>